pyproject.toml
[tool.poetry]
name = "tcc-pos-puc"
version = "0.1.0"
description = ""
authors = ["José Nunes <josenunesra@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
confluent-kafka = "^1.8.2"
psycopg2-binary = "^2.9.3"
apache-airflow = "^2.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
